Move overlayInstance.ts
