+++ packages/core/src/components/overlay/overlayProps.ts
@@ -15,7 +15,7 @@
  */
 
 import type { Props } from "../../common/props";
-import type { OverlayInstance } from "../overlay2/overlayInstance";
+import type { OverlayInstance } from "./overlayInstance";
 
 export interface OverlayableProps extends OverlayLifecycleProps {
     /**
